requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
django-environ==0.11.2
redis==5.0.8
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
drf-spectacular==0.27.2
